Fix multiple arrows in same tick
If two (or more) arrows with different amplifiers land in the same tick, only is deleted, but both summon their correspondig creeper. This results in two creepers from one of the arrows.
If two (or more) arrows with the same amplifier land in the same tick, one is processed with a one tick delay. This results in two target selections instead of one.